crypto_bot/requirements.txt
click
flake8
numpy
pandas
psycopg2-binary
pytest
pytest-mock
pytz
requests
setuptools
-e git+https://github.com/dominikheinisch/pybitbay@v0.0.1#egg=pybitbay